:alien: Remove isFlatSpecimenModel flag
