+++ config/output/config/qa-include.conf
@@ -1273,7 +1273,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1321,7 +1321,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1369,7 +1369,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1417,7 +1417,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1465,7 +1465,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1513,7 +1513,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1561,7 +1561,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1609,7 +1609,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1657,7 +1657,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1705,7 +1705,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1753,7 +1753,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1801,7 +1801,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1846,7 +1846,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             table {
@@ -1872,6 +1872,10 @@ datalake {
             ]
             path="/normalized/sequencing_experiment"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             table {
                 database="include_portal_qa"
@@ -1896,6 +1900,10 @@ datalake {
             ]
             path="/normalized/sequencing_experiment_genomic_file"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             table {
                 database="include_portal_qa"
@@ -1920,6 +1928,10 @@ datalake {
             ]
             path="/normalized/sequencing_center"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             table {
                 database="include_portal_qa"
@@ -2009,69 +2021,14 @@ datalake {
             ]
             path="/enriched/histology_disease"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             table {
                 database="include_portal_qa"
                 name="histology_disease"
-            }
-            writeoptions {
-                "created_on_column"="created_on"
-                "is_current_column"="is_current"
-                overwriteSchema="true"
-                "updated_on_column"="updated_on"
-                "valid_from_column"="valid_from"
-                "valid_to_column"="valid_to"
-            }
-        },
-        {
-            format=DELTA
-            id="enriched_specimen"
-            keys=[]
-            loadtype=OverWritePartition
-            partitionby=[
-                "study_id"
-            ]
-            path="/enriched/specimen"
-            readoptions {}
-            storageid=storage
-            table {
-                database="include_portal_qa"
-                name="enriched_specimen"
-            }
-            writeoptions {
-                "created_on_column"="created_on"
-                "is_current_column"="is_current"
-                overwriteSchema="true"
-                "updated_on_column"="updated_on"
-                "valid_from_column"="valid_from"
-                "valid_to_column"="valid_to"
-            }
-        },
-        {
-            format=DELTA
-            id="normalized_snv"
-            keys=[]
-            loadtype=OverWritePartition
-            partitionby=[
-                "study_id",
-                "has_alt",
-                chromosome
-            ]
-            path="/normalized/snv"
-            readoptions {}
-            repartition {
-                column-names=[
-                    chromosome,
-                    start
-                ]
-                kind=RepartitionByRange
-                n=100
-                sort-columns=[]
-            }
-            storageid=storage
-            table {
-                database="include_portal_qa"
-                name="normalized_snv"
             }
             writeoptions {
                 "created_on_column"="created_on"
@@ -2116,10 +2073,49 @@ datalake {
             ]
             path="/enriched/family"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             table {
                 database="include_portal_qa"
                 name="enriched_family"
+            }
+            writeoptions {
+                "created_on_column"="created_on"
+                "is_current_column"="is_current"
+                overwriteSchema="true"
+                "updated_on_column"="updated_on"
+                "valid_from_column"="valid_from"
+                "valid_to_column"="valid_to"
+            }
+        },
+        {
+            format=DELTA
+            id="normalized_snv"
+            keys=[]
+            loadtype=OverWritePartition
+            partitionby=[
+                "study_id",
+                "has_alt",
+                chromosome
+            ]
+            path="/normalized/snv"
+            readoptions {}
+            repartition {
+                column-names=[
+                    chromosome,
+                    start
+                ]
+                kind=RepartitionByRange
+                n=100
+                sort-columns=[]
+            }
+            storageid=storage
+            table {
+                database="include_portal_qa"
+                name="normalized_snv"
             }
             writeoptions {
                 "created_on_column"="created_on"
@@ -2266,4 +2262,3 @@ datalake {
     ]
 }
 "dataservice_url"="https://kf-api-dataservice-qa.kidsfirstdrc.org"
-isFlatSpecimenModel=false
