+++ config/output/config/ucsf.conf
@@ -1145,7 +1145,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1189,7 +1189,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1233,7 +1233,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1277,7 +1277,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1321,7 +1321,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1365,7 +1365,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1409,7 +1409,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1453,7 +1453,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1497,7 +1497,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1541,7 +1541,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1585,7 +1585,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1629,7 +1629,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1670,7 +1670,7 @@ datalake {
             readoptions {}
             repartition {
                 kind=Coalesce
-                n=10
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1692,6 +1692,10 @@ datalake {
             ]
             path="/normalized/sequencing_experiment"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             writeoptions {
                 "created_on_column"="created_on"
@@ -1712,6 +1716,10 @@ datalake {
             ]
             path="/normalized/sequencing_experiment_genomic_file"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             writeoptions {
                 "created_on_column"="created_on"
@@ -1732,6 +1740,10 @@ datalake {
             ]
             path="/normalized/sequencing_center"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
             storageid=storage
             writeoptions {
                 "created_on_column"="created_on"
@@ -1809,56 +1821,9 @@ datalake {
             ]
             path="/enriched/histology_disease"
             readoptions {}
-            storageid=storage
-            writeoptions {
-                "created_on_column"="created_on"
-                "is_current_column"="is_current"
-                overwriteSchema="true"
-                "updated_on_column"="updated_on"
-                "valid_from_column"="valid_from"
-                "valid_to_column"="valid_to"
-            }
-        },
-        {
-            format=DELTA
-            id="enriched_specimen"
-            keys=[]
-            loadtype=OverWritePartition
-            partitionby=[
-                "study_id"
-            ]
-            path="/enriched/specimen"
-            readoptions {}
-            storageid=storage
-            writeoptions {
-                "created_on_column"="created_on"
-                "is_current_column"="is_current"
-                overwriteSchema="true"
-                "updated_on_column"="updated_on"
-                "valid_from_column"="valid_from"
-                "valid_to_column"="valid_to"
-            }
-        },
-        {
-            format=DELTA
-            id="normalized_snv"
-            keys=[]
-            loadtype=OverWritePartition
-            partitionby=[
-                "study_id",
-                "has_alt",
-                chromosome
-            ]
-            path="/normalized/snv"
-            readoptions {}
             repartition {
-                column-names=[
-                    chromosome,
-                    start
-                ]
-                kind=RepartitionByRange
-                n=100
-                sort-columns=[]
+                kind=Coalesce
+                n=1
             }
             storageid=storage
             writeoptions {
@@ -1900,6 +1865,41 @@ datalake {
             ]
             path="/enriched/family"
             readoptions {}
+            repartition {
+                kind=Coalesce
+                n=1
+            }
+            storageid=storage
+            writeoptions {
+                "created_on_column"="created_on"
+                "is_current_column"="is_current"
+                overwriteSchema="true"
+                "updated_on_column"="updated_on"
+                "valid_from_column"="valid_from"
+                "valid_to_column"="valid_to"
+            }
+        },
+        {
+            format=DELTA
+            id="normalized_snv"
+            keys=[]
+            loadtype=OverWritePartition
+            partitionby=[
+                "study_id",
+                "has_alt",
+                chromosome
+            ]
+            path="/normalized/snv"
+            readoptions {}
+            repartition {
+                column-names=[
+                    chromosome,
+                    start
+                ]
+                kind=RepartitionByRange
+                n=100
+                sort-columns=[]
+            }
             storageid=storage
             writeoptions {
                 "created_on_column"="created_on"
@@ -2030,4 +2030,3 @@ datalake {
     ]
 }
 "dataservice_url"=""
-isFlatSpecimenModel=true
