+++ config/src/main/scala/bio/ferlab/fhir/etl/config/ConfigurationGenerator.scala
@@ -22,8 +22,6 @@ object ConfigurationGenerator extends App {
     sources.map(ds => ds.copy(table = ds.table.map(t => TableConf(database, t.name))))
   }
 
-  def isFlatSpecimenModel(project: String): Boolean = project == pKfStrides
-
   private val partitionByStudyId = List("study_id")
   val sourceNames: Seq[SourceConfig] = Seq(
     SourceConfig("family_relationship", partitionByStudyId),
@@ -198,7 +196,7 @@ object ConfigurationGenerator extends App {
     "spark.databricks.delta.replaceWhere.constraintCheck.enabled" -> "false"
   )
   conf.foreach { case (project, _) =>
-    ConfigurationWriter.writeTo(s"config/output/config/dev-$project.conf", ETLConfiguration(isFlatSpecimenModel(project), DatalakeConf(
+    ConfigurationWriter.writeTo(s"config/output/config/dev-$project.conf", ETLConfiguration(DatalakeConf(
       storages = List(
         StorageConf(storage, "s3a://storage", S3),
         gnomadStorage
@@ -227,7 +225,7 @@ object ConfigurationGenerator extends App {
     ))
 
 
-    ConfigurationWriter.writeTo(s"config/output/config/qa-$project.conf", ETLConfiguration(isFlatSpecimenModel(project), DatalakeConf(
+    ConfigurationWriter.writeTo(s"config/output/config/qa-$project.conf", ETLConfiguration(DatalakeConf(
       storages = List(
         StorageConf(storage, s"s3a://${conf(project)("bucketNamePrefix")}-qa", S3),
         gnomadStorage
@@ -239,7 +237,7 @@ object ConfigurationGenerator extends App {
       dataservice_url = "https://kf-api-dataservice-qa.kidsfirstdrc.org"
     ))
 
-    ConfigurationWriter.writeTo(s"config/output/config/prd-$project.conf", ETLConfiguration(isFlatSpecimenModel(project), DatalakeConf(
+    ConfigurationWriter.writeTo(s"config/output/config/prd-$project.conf", ETLConfiguration(DatalakeConf(
       storages = List(
         StorageConf(storage, s"s3a://${conf(project)("bucketNamePrefix")}-prd", S3),
         gnomadStorage
@@ -252,7 +250,7 @@ object ConfigurationGenerator extends App {
     ))
   }
 
-  ConfigurationWriter.writeTo(s"config/output/config/ucsf.conf", ETLConfiguration(isFlatSpecimenModel = true, DatalakeConf(
+  ConfigurationWriter.writeTo(s"config/output/config/ucsf.conf", ETLConfiguration(DatalakeConf(
     storages = List(
       StorageConf(storage, s"s3a://d3b-portal-65-4-r-us-west-2.sec.ucsf.edu", S3),
       gnomadStorage
